% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PokemonPackage.R
\name{atkCheck}
\alias{atkCheck}
\title{Function that takes in two pokemon and returns true if the attack stat of the first pokemon is higher}
\usage{
atkCheck(attacker, defender)
}
\arguments{
\item{attacker}{name of the attacking pokemon}

\item{defender}{name of the defending pokemon}
}
\value{
None
}
\description{
Function that takes in two pokemon and returns true if the attack stat of the first pokemon is higher
}
\examples{
atkCheck("Charmander","Urshifu" )
}
