% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PokemonPackage.R
\name{pokeBall}
\alias{pokeBall}
\title{Function to see if you catch the pokemon or not}
\usage{
pokeBall(pokemon, HPcurrent)
}
\arguments{
\item{pokemon}{name of the pokemon we want to catch}

\item{HPcurrent}{current health of pokemon}
}
\value{
None
}
\description{
Function to see if you catch the pokemon or not
}
\examples{
HPcurrent("Charmander", 20 )
}
