% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PokemonPackage.R
\name{pokeDex}
\alias{pokeDex}
\title{Function for getting data about a single pokemon}
\usage{
pokeDex(pokemon)
}
\arguments{
\item{pokemon}{name of the pokemon we want information about}
}
\value{
None
}
\description{
Function for getting data about a single pokemon
}
\examples{
pokeDex("Charmander")
pokeDex("Litwick")
}
