% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PokemonPackage.R
\name{pokemon}
\alias{pokemon}
\title{R6 Class representing a pokemon}
\description{
R6 Class representing a pokemon

R6 Class representing a pokemon
}
\details{
A pokemon has a name, two types, and HP.
}
\examples{

## ------------------------------------------------
## Method `pokemon$show`
## ------------------------------------------------

myPokemon = new$pokemon("Squirtle")
myPokemon$show()

## ------------------------------------------------
## Method `pokemon$changeHP`
## ------------------------------------------------

myPokemon = new$pokemon("Squirtle")
myPokemon$changeHP(5)

## ------------------------------------------------
## Method `pokemon$fight`
## ------------------------------------------------

myPokemon = new$pokemon("Squirtle")
myPokemon2 = new$pokemon("Charmander")
myPokemon$fight(myPokemon2)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the pokemon}

\item{\code{type1}}{The primary type of the pokemon}

\item{\code{type2}}{The secondary type of the pokemon}

\item{\code{startingHP}}{The starting HP of the pokemon}

\item{\code{awake}}{The awake status of the pokemon}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pokemon$new()}}
\item \href{#method-show}{\code{pokemon$show()}}
\item \href{#method-changeHP}{\code{pokemon$changeHP()}}
\item \href{#method-fight}{\code{pokemon$fight()}}
\item \href{#method-clone}{\code{pokemon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new pokemon object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pokemon$new(name = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{pokemon} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
Show the pokemon and it's attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pokemon$show()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myPokemon = new$pokemon("Squirtle")
myPokemon$show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-changeHP"></a>}}
\if{latex}{\out{\hypertarget{method-changeHP}{}}}
\subsection{Method \code{changeHP()}}{
Change the HP of the pokemon by a user provided value, updates awake if HP<=0
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pokemon$changeHP(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Amount to change the HP by (positive or negative)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myPokemon = new$pokemon("Squirtle")
myPokemon$changeHP(5)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fight"></a>}}
\if{latex}{\out{\hypertarget{method-fight}{}}}
\subsection{Method \code{fight()}}{
Takes in another pokemon, has the two pokemon 'fight', and randomly decreases their HP
If HP<=0, awake is set to FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pokemon$fight(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{A \code{pokemon} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myPokemon = new$pokemon("Squirtle")
myPokemon2 = new$pokemon("Charmander")
myPokemon$fight(myPokemon2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pokemon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
