% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PokemonPackage.R
\docType{data}
\name{pokemon_data}
\alias{pokemon_data}
\title{Data about Pokemon
a dataset about pokemon from source}
\format{
Pokemon dataframe with 1,076 entries, 47 columns
}
\source{
Kaggle
}
\usage{
pokemon_data
}
\description{
Data about Pokemon
a dataset about pokemon from source
}
\keyword{datasets}
