% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PokemonPackage.R
\name{atkCheck}
\alias{atkCheck}
\title{Function for battling pokemon}
\usage{
atkCheck(attacker, defender)
}
\arguments{
\item{attacker}{name of the attacking pokemon}

\item{defender}{name of the defending pokemon}
}
\value{
None
}
\description{
Function for battling pokemon
}
\examples{
atkCheck("Charmander","Urshifu" )
}
