% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PokemonPackage.R
\name{Pokemon}
\alias{Pokemon}
\title{R6 Class representing a Pokemon}
\description{
R6 Class representing a Pokemon

R6 Class representing a Pokemon
}
\details{
A pokemon has a name, two types, and HP.
}
\examples{

## ------------------------------------------------
## Method `Pokemon$show`
## ------------------------------------------------

p = Pokemon$new("Charmander")
p$show()

## ------------------------------------------------
## Method `Pokemon$changeHP`
## ------------------------------------------------

p = new$Pokemon("Charmander")
p$changeHP(3)

## ------------------------------------------------
## Method `Pokemon$fight`
## ------------------------------------------------

p = Pokemon$new("Charmander")
opponent = Pokemon$new("Squirtle")
p$fight(opponent)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the Pokemon.}

\item{\code{type1}}{The primary type of Pokemon.}

\item{\code{type2}}{The secondary type of Pokemon.}

\item{\code{starting_HP}}{The starting number of HP points.}

\item{\code{updated_HP}}{The updated number of HP points after the fight function is run.}

\item{\code{awake}}{Whether or not the Pokemon is awake. The Pokemon is awake if the HP is greater than 0.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Pokemon$new()}}
\item \href{#method-show}{\code{Pokemon$show()}}
\item \href{#method-changeHP}{\code{Pokemon$changeHP()}}
\item \href{#method-fight}{\code{Pokemon$fight()}}
\item \href{#method-clone}{\code{Pokemon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Pokemon. Return the Pokemon name, primary type, secondary type, starting HP, and updated HP.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$new(name = NA)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Pokemon} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
Give information about the Pokemon in list form.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$show()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{p = Pokemon$new("Charmander")
p$show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-changeHP"></a>}}
\if{latex}{\out{\hypertarget{method-changeHP}{}}}
\subsection{Method \code{changeHP()}}{
Change the Pokemon HP based on input from the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$changeHP(c)}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{p = new$Pokemon("Charmander")
p$changeHP(3)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fight"></a>}}
\if{latex}{\out{\hypertarget{method-fight}{}}}
\subsection{Method \code{fight()}}{
Take two pokemon and make them fight. Both Pokemon will lose 20 points per fight.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$fight(opponent)}\if{html}{\out{</div>}}
}

\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{p = Pokemon$new("Charmander")
opponent = Pokemon$new("Squirtle")
p$fight(opponent)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
