% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PokemonPackage.R
\docType{data}
\name{pokemon_data}
\alias{pokemon_data}
\title{Pokemon Data}
\format{
Each observation represents a different Pokemon. The observation includes the Pokemon name, type1, type2, starting HP, and whether or not they are awake.
}
\source{
: kaggle.com
}
\usage{
pokemon_data
}
\description{
Pokemon Data
}
\keyword{datasets}
